chore: jenny_game 컴포넌트 분리 + 각 페이지 렌더링되는 버튼 수정 + bgm play 로직 수정
